djongo1/requirements.txt
appdirs==1.4.3
astroid==2.0.4
attrs==18.2.0
black==18.9b0
certifi==2020.6.20
Click==7.0
csv2arff==2.0
Django==2.1.15
djangorestframework==3.11.0
djongo==1.3.3
dnspython==1.16.0
graphviz==0.8.4
idna==2.6
install==1.3.3

lazy-object-proxy==1.3.1
mccabe==0.6.1
mxnet==1.5.0
onnx==1.7.0
pylint==2.1.1
pymongo==3.10.1
python-dotenv==0.14.0
python-weka-wrapper3==0.1.11
pytz==2020.1
sqlparse==0.2.4
toml==0.10.0
typing-extensions==3.7.4.2
wincertstore==0.2
wrapt==1.12.1
